feat(qtum): implement delegation ops
